added MapObjectType enum and Array of MapObjectType in MapService
