+++ core/src/com/geometric/wars/maps/MapObjectType.java
@@ -1,0 +1,11 @@
+package com.geometric.wars.maps;
+
+public enum MapObjectType {
+    EMPTY,
+    WALL,
+    PLAYER,
+    /**
+     * map object that represents cell where player already has been but now is moving to different cell
+     */
+    PLAYERSPECTRUM;
+}
