+++ core/src/com/geometric/wars/player/ShooterPlayersController.java
@@ -1,7 +1,5 @@
 package com.geometric.wars.player;
 
-import com.geometric.wars.maps.MapObjectCheckerService;
-
 public class ShooterPlayersController extends PlayersController {
 
     private int healthPoints;
